Made the Keyboard Smaller
In the Riddle when you have the keyboard open I left some room on top
to it gives you another option to get out of the keyboard. You can
click back on the riddle space. We can actually make the answer list as
small as we want. What ever ends up being the best usability wise.